HP animation
The damage and total HP have been temporarily inflated to allow better
testing of the animation speed.

+++ include/battle/Pokemon.hpp
@@ -76,7 +76,7 @@ struct Pokemon {
     int asleepRounds;
     int level;
     std::array<int, 6> stats;
-    int currentHP;
+    float currentHP;
 };
 
 #endif
